Initial Commit
Going to experiment building DSP plugins in C++!

Going to build with JUCE, it seems to be the framework that people use for cross-DAW plugins, rather than rolling your own implementation of that cross-compatibility logic.
https://github.com/juce-framework/JUCE

Heard about SketchCassette II, that was the inspiration for looking into this. I saw that they develop for the three main plugin formats, I was curious how one would do that.
https://aberrantdsp.com/plugins/sketchcassette/

This is another crossover of learning for me, learning how to use plugins and mixing inside the DAW with my own music creation process, but also how to learn how to build those myself too, and how those work.

Glad to have a need for C++ now, been wanting to have it as part of my toolbelt for a while, but haven't had the 'need' to use it, per se. That seems to be the drive for me to learn something new, it needs to be something to fill the void of what I need. This sounds very similar to a video I saw of Linux Torvalds about his projects too.

https://www.youtube.com/watch?v=jJEiPQyKH4E

The Fluke, Nobody's Here, Tiny Tears, Earth Day (DT)