About the Project!
Awesome! Managed to build the project in Xcode, and it automatically put it inside of the place Logic looks for plugins. It runs perfect as well. Now I want to learn how to do that without needing Xcode, hopefully. I know it might not be possible. We'll see! At least it works at all, I'd rather that be the case.

https://www.youtube.com/watch?v=PltjGej4Jes

Opening it in Xcode made a bunch of config files. I didn't add them here because I don't think they're quite necessary. I don't know for sure though. Man, it would be great to just use the terminal toolchain and or VSCode. Maybe I should use xcodebuild instead of cmake? That might be the issue here. I'll look into it. I thought I heard that mentioned somewhere previously. Not sure if it was to do with JUCE specifically.